# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Campus Skill Share API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      tags: [users]
      operationId: getMyProfile
      summary: Get current user profile (including role and display name)
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [users]
      operationId: updateMyProfile
      summary: Update current user display name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /posts:
    get:
      tags: [posts]
      operationId: listPosts
      summary: List skill posts (newest first)
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListResponse'
    post:
      tags: [posts]
      operationId: createPost
      summary: Create a new skill post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /posts/{postId}:
    get:
      tags: [posts]
      operationId: getPost
      summary: Get a single skill post by ID
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [posts]
      operationId: deletePost
      summary: Delete a post (admin or post owner)
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /messages/{otherUserId}:
    get:
      tags: [messages]
      operationId: getConversation
      summary: Get messages between current user and another user
      parameters:
        - name: otherUserId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /messages:
    post:
      tags: [messages]
      operationId: sendMessage
      summary: Send a message to another user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageInput'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /messages/conversations:
    get:
      tags: [messages]
      operationId: listConversations
      summary: List all conversations for the current user
      responses:
        '200':
          description: Conversation summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationListResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/posts:
    get:
      tags: [admin]
      operationId: adminListPosts
      summary: Admin - list all posts
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: All posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListResponse'
        '403':
          description: Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  schemas:
...
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UserProfile:
      type: object
      required: [id, email, displayName, role, profileImageUrl, createdAt]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        displayName:
          type: string
        role:
          type: string
          enum: [student, admin]
        profileImageUrl:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    UserProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
          minLength: 2
    PostAuthor:
      type: object
      required: [id, displayName, profileImageUrl]
      properties:
        id:
          type: string
        displayName:
          type: string
        profileImageUrl:
          type: [string, 'null']
    Post:
      type: object
      required: [id, title, category, description, createdAt, author]
      properties:
        id:
          type: string
        title:
          type: string
        category:
          type: string
        description:
          type: string
        availability:
          type: [string, 'null']
        priceRate:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        author:
          $ref: '#/components/schemas/PostAuthor'
    PostInput:
      type: object
      required: [title, category, description]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 150
        category:
          type: string
          enum: [Tutoring, Design, Music, Tech, Language, Other]
        description:
          type: string
          minLength: 20
          maxLength: 500
        availability:
          type: string
          maxLength: 200
        priceRate:
          type: string
          maxLength: 100
    PostListResponse:
      type: object
      required: [posts, total]
      properties:
        posts:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        total:
          type: integer
...
    Message:
      type: object
      required: [id, senderId, receiverId, content, createdAt]
      properties:
        id:
          type: string
        senderId:
          type: string
        receiverId:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    MessageInput:
      type: object
      required: [receiverId, content]
      properties:
        receiverId:
          type: string
        content:
          type: string
          minLength: 1
          maxLength: 2000
    MessageListResponse:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
    ConversationSummary:
      type: object
      required: [otherUser, lastMessage, unreadCount]
      properties:
        otherUser:
          $ref: '#/components/schemas/PostAuthor'
        lastMessage:
          $ref: '#/components/schemas/Message'
        unreadCount:
          type: integer
    ConversationListResponse:
      type: object
      required: [conversations]
      properties:
        conversations:
          type: array
          items:
            $ref: '#/components/schemas/ConversationSummary'
...