# lib/api-zod/src/generated/types/conversationListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Campus Skill Share API
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationSummary } from './conversationSummary';

export interface ConversationListResponse {
  conversations: ConversationSummary[];
}