# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Campus Skill Share API
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  receiverId: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  content: string;
}