# lib/api-zod/src/generated/types/userProfileUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Campus Skill Share API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileUpdate {
  /** @minLength 2 */
  displayName?: string;
}